fix(dashboard): keep quality ratchet within file baselines
